Increase default value of net.ipv4.igmp_max_memberships
The linux kernel default for igmp_max_memberships is 20 [0].

For vxlan project networks, it is possible to specify a range
of multicast addresses for use with VXLAN tunnel broadcast traffic.
In the case where a range is specified a unique multicast group
is assigned by neutron to each vxlan network, and once
igmp_max_memberships is reached for a host, vxlan interfaces will
silently fail to join the multicast group.

This results in instances being unable to DHCP or ARP, and errors
in the neutron agent log [1].

This patch increases the default value to 1024 which should be
sufficient for all but the very largest deployments.

[0] https://sysctl-explorer.net/net/ipv4/igmp_max_memberships/
[1] http://paste.openstack.org/show/796504

Change-Id: I24a0ca3001098b278d341f710ca2e4f00dc3e09a
(cherry picked from commit 7cbf51170c9ad25ac6a36ebc94561018a3b310b5)